simplify primary inspect_nf script
